New subcommand and probably bad changes to `Database`
The `init` subcommand has been added which just calls `Database::init()`.

Changed the most of `Database`'s' methods, so that they return `Self` (
whether wrapped in a `Result` or not), which means that they can be
chained like with the builder pattern. Probably a bad idea but that's a
problem for future me :^)

Also made an existing database not cause `Database::init()` to `bail!`,
and instead print a nice message to the console and exit.

Signed-off-by: Erick Howard <needlesslygrim@proton.me>