Blazingly Fast web api :^)
Uses `tiny_http`. Currently, it only serves `index.{html, css, js}`, and
has an api for querying logins. It also doesn't actually sync the
datbase in serve mode since it doesn't have a way of shutting down
properly. A new subcommand is introduced, `serve`, which runs the web
server. The web-capable version of safe is behind a feature flag, which
allows it to be smaller in case the web interface is unneeded.

It's extremely fast, a request query of 'msft' in a database of ~35
million logins can be completed in around 15 seconds, with ~6 seconds of
transfer time.

- Also fixed up the formatting in `Cargo.toml`
- Renamed `Database::query()` to `Database::query_interactive()`
- Possibly other things I've forgotten about since doing them

Signed-off-by: Erick Howard <needlesslygrim@proton.me>