Implemented hot-reloading for the web interface
In debug mode, instead of `include_bytes!()`ing the web interface, we
will instead, just open the file in question each time it's requested,
and serve what what we get when we read it. Behold! Hot reloading.

Also changed the path to `/index.js` to `$PROJECT_ROOT/dist/index.js`,
since Typescript is now in use, and we need a compile step.

Signed-off-by: Erick Howard <needlesslygrim@proton.me>