Implement an ID system for the database
Each login now has an associated UUID which is how they are removed from
the database. Because we now have uniique indices, we can use a hashmap,
which will increase our removal performance etc. HHowever, overall it
might be slower, so I might have to look at the performance difference
between using a `Vec<Login>` with the UUID inside the `Login`, and using
a `Hashmap<Uuid, Login>`.

I also removed the support for `fuzzy_matcher`, because it seemed to be
about the same as `nucleo_matcher` speedwise (with very unscientific
benchmarks), but having them both wasn't much use.

Lastly, the the removal API is now implemented properly, since we have a
proper ID system :^).

Signed-off-by: Erick Howard <needlesslygrim@proton.me>