Add config file support, change the db format, and add graceful shutdown
This commit brings three changes, the first is the addition
of configuration file support, which currently only allows for
configuration of the webserver port location, so honestly rather useless
and not worth the hassle that it was to implement.

The second is using MessagePack instead of RON to store the database.
There wasn't actually much reason to do this to be honest, but it might
save space?

Both these files are now stored in the proper places according to the
platform. The config file is stored in the config file dir, e.g.
`~/.config/gondolin/` on Linux (assuming sane XDG configuration), and
the database is stored in the "data dir" that `directories` provides,
which is `~/.local/share/gondolin/` on Linux.

The third, and arguably only useful change, is enabling a sort of
graceful shutdown for the web server. When it recives SIGINT, it will
synchronise the database file to disk. The minor problem with it now
is that only after each request can we actually check whether we should
shutdown or not. This is because we run it at the end of the event loop
that iterates through `server.incoming_requests()`, which blocks while
waiting for a request. The alternative is to just `loop`, and check
`server.incoming_requests().next()` ourselves. The problem with *that*
approach is that it uses 6% of my CPU constantly, since it's just an
infinite loop.

Signed-off-by: Erick Howard <needlesslygrim@proton.me>