First version of the web interface! :^)
This commit implements the first version of the web interface which has
been in progress for about a month. All styling is done with Tailwind,
for I do not know (read: cannot be bothered to learn at the moment) CSS,
so Tailwind allows me to do some reasonably advanced stuff. It supports
both dark and light themes, which is nice. It also supports different
screen sizes to an extent. Sizing with HTML and CSS is still a little
confusing, so in some places, e.g. the searchbar, things are hardcoded
so that they sort-of work. Both pages which are currently implemented
require 0 JS, which I see as a win. However in subsequent versions JS
may be required to show dialogs.

The query page is rendered server-side, using a hacky method to achieve
what is essentially HTML templating in Rust, using format strings.
This is probably unperformant, and doesn't support hot-reloading as
the front- page does, but it works for now, and means I don't need to
introduce yet another dependancy.

A problem with the current implementation of everything, is that we
have no component framework of any kind, which is annoying not only for
use with Tailwind, and will be worse once I need to implement dialogs.
For this reason I may need to introduce something like JSX (if that can
be used without React), so that I don't need to read the 200 chars of
Tailwind classes per line that I currently do when reading any of the
HTML I've written.

Another interesting detail is that the packages are managed with Bun (
at least I think so, I'm not really sure how JS package managers work
:þ). This is not for any real reason, apart from the Bun package manager
being extremely quick. If anyone should be reading these commit logs and
they see a potential problem with this, please inform me :^).

That should be everything, apart from one small change I made to
`lib.rs` which removes an unnessecary comment.

Signed-off-by: Erick Howard <needlesslygrim@proton.me>