Slight improvement to query UI and searchbox
The grid of results is now a lot better. I rewrote the card component
so that it is a lot less complex, and added a button to the bottom
which will be hooked up to a function to remove that login soon. The
responsive design will also be of better quality, and there won't be any
overlap issues.

The searchbar just had its placeholder text set, as well as the `aria-
label`. Other accessibility work has yet to be done, but it's a good
first step.

Signed-off-by: Erick Howard <needlesslygrim@proton.me>