Add support for removing logins through the web interface
The delete button at the bottom of each card (I think that was added in
the previous commit?) now actually deletes the login. It will first ask
the user if they are sure they want to delete the login, and if so, will
delete it. Should it be deleted successfully, it will show a success
message if the two following conditions are met. First, if the message
has never before. Second, if the value of `show-remove-success` in Local
Storage is set to `"true"`. I will probably remove this at some point,
as according to MDN showing confirm messages is largely pointless, and
should be avoided.

Also:
- Prettier config
- Format README and TODO
- Add one TODO
- Fix `tsconfig.json`

Signed-off-by: Erick Howard <needlesslygrim@proton.me>