Big change, added a form to add new logins to the database
Honestly, it's too late for me to remeber everything that has changed
(it's 00:38), so this commit message isn't going to be very useful.
Basically I added a form, and added support for that in `net.rs`.
Fixed other things like the scripts being sent with a mimetype of
`application/json` incorrectly.

Signed-off-by: Erick Howard <needlesslygrim@proton.me>